requirements.txt
asgiref==3.3.1
astroid==2.4.2
beautifulsoup4==4.9.3
Django==3.1.4
django-bootstrap4==2.3.1
django-bulma==0.6.0
django-filter==2.4.0
django-pdfkit==0.3.1
django-summernote==0.8.11.6
django-taggit==1.3.0
gunicorn==20.0.4
hashtag==0.1
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
pdfkit==0.6.0
Pillow==8.0.1
psycopg2-binary==2.8.6
pylint==2.6.0
pytz==2020.4
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
toml==0.10.2
whitenoise==5.2.0
wrapt==1.12.1
